pyproject.toml
[project]
name = "evals"
version = "0.1.0"
description = "Swissai evaluation scripts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "iso639-lang>=2.6.1",
    "pandas>=2.3.0",
    "prtpy>=0.8.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "wandb>=0.20.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
